requirements.txt
pygame
numpy
scipy
pyopengl
pyopengl_accelerate
fonttools
